+++ lib/error-check.sh
@@ -17,6 +17,41 @@ function errorChk() {
 
 function deployChk() {
 	if [[ "${DEPLOY}" == *"mina"* ]]; then
+		# Get variables organized
+		grep -n "user" "${WORKPATH}/${APP}"/config/deploy.rb > "${trshFile}"
+		MINAUSER=$(awk -F\' '{print $2,$4}' ${trshFile})
+		echo -n "${MINAUSER}" > "${statFile}"
+		echo -n "@" >> ${statFile}
+		grep -n "domain" "${WORKPATH}/${APP}"/config/deploy.rb > "${trshFile}"
+		MINADOMAIN=$(awk -F\' '{print $2,$4}' ${trshFile})
+		echo -n "${MINADOMAIN}" >> "${statFile}"
+		SSHTARGET=$(sed -r 's/\s+//g' ${statFile})
+		
+		# SSH check
+		trace "Testing connection for ${SSHTARGET}"
+		SSHSTATUS=$(ssh -o BatchMode=yes -o ConnectTimeout=10 ${SSHTARGET} echo ok 2>&1)
+
+		if [[ "${SSHSTATUS}" == *"ok"* ]] ; then
+			# Continue deploying
+			trace "OK"
+		elif [[ "${SSHSTATUS}" == *"Permission denied"* ]] ; then
+			# Not authorized, no key etc.
+			error "Connection refused for ${SSHTARGET}"
+		else
+			warning "Connection for ${SSHTARGET} not established, an unknown error occurred."
+			# Ok now re-run mina in verbose mode if someone is at the console
+
+			# If FORCE=1 then simply exit
+			if [[ "${FORCE}" = "1" ]] || yesno --default yes "Retry ${DEPLOY} in verbose mode? [Y/n] "; then
+				eval "${DEPLOY}" | tee --append "${logFile}"
+			else
+				error_msg="Connection for ${SSHTARGET} not established, an unknown error occurred."
+				errorExit
+			fi
+		fi
+
+		# Try to --simulate the command
+		trace "Testing deployment command: ${DEPLOY}"
 		mina --simulate deploy &>> /dev/null
 		EXITCODE=$?; 
 		if [[ "${EXITCODE}" != 0 ]]; then 
@@ -24,6 +59,7 @@ function deployChk() {
 			error_msg="Deployment exited due to a configuration problem (Error ${EXITCODE})"
 			errorExit
 		fi
+		trace "OK"
 	fi
 }