+++ lib/ssh-check.sh
@@ -3,10 +3,11 @@
 # sshChk()
 #
 # 
-trace "Loading SSH checks"
+#trace "Loading SSH checks"
 
 function sshChk() {
 	notice "Checking SSH Keys..."
 	ssh -oStrictHostKeyChecking=no git@bitbucket.org
 	ssh -oStrictHostKeyChecking=no git@github.org
 }
+